Add authorization for card modifications and improve db transactions

+++ app/Http/Controllers/CardController.php
@@ -7,11 +7,13 @@ use App\Http\Services\CardService;
 use App\Models\Card;
 use App\Models\Listing;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
 class CardController extends Controller
 {
+    
     /**
      * Display a listing of the resource.
      *
@@ -19,10 +21,11 @@ class CardController extends Controller
      */
     public function index(Request $request)
     {
+        $listingId = $request->query('listing_id');
+        $this->authorize('viewAny', [Card::class, $listingId]);
         $request->validate([
             'listing_id' => 'required|exists:listings,id'
         ]);
-        $listingId = $request->query('listing_id');
         $cards = Card::where('listing_id', $listingId)->get();
         return $cards;
     }
@@ -40,6 +43,7 @@ class CardController extends Controller
             'title' => 'required|min:10|max:200',
             'description' => 'max:2000',
         ]);
+        $this->authorize('create', [Card::class, $data['listing_id']]);
         $listing = Listing::findOrFail($request->listing_id);
         $data['index'] = $listing->cards()->max('index') === null ? 0 : $listing->cards()->max('index') + 1;
         $card = new Card($data);
@@ -56,6 +60,7 @@ class CardController extends Controller
     public function show($id)
     {
         $card = Card::findOrFail($id);
+        $this->authorize('view', [Card::class, $card->listing_id]);
         return $card;
     }
 
@@ -74,6 +79,7 @@ class CardController extends Controller
             'description' => 'max:2000'
         ]);
         $card = Card::findOrFail($id);
+        $this->authorize('update', [$card, $data['listing_id']]);
         $card->update($data);
         return $card;
     }
@@ -86,22 +92,26 @@ class CardController extends Controller
             'current_listing_index' => 'required|integer',
         ]);
         try{
-            $card = null;
-            DB::transaction(function() use(&$card, $data, $id){
-                $card = Card::findOrFail($id);
-                $boardId = $card->listing->board->id;
-                $previousListingId = isset($data['previous_listing_index']) ? Listing::where([['board_id', $boardId], ['index', $data['previous_listing_index']]])->first()->id : null;
-                $currentListingId = Listing::where([['board_id', $boardId], ['index', $data['current_listing_index']]])->first()->id;
-                $card->update([
-                    'index' => $data['current_index'],
-                    'listing_id' => $currentListingId,
-                ]);
-                if($previousListingId)
-                    CardService::resetIndexesAfterAction($previousListingId);
-                CardService::resetIndexesAfterAction($currentListingId);
-            });
+            DB::beginTransaction();
+            $card = Card::findOrFail($id);
+            $boardId = $card->listing->board->id;
+            $previousListingId = isset($data['previous_listing_index']) ? Listing::where([['board_id', $boardId], ['index', $data['previous_listing_index']]])->first()->id : null;
+            $currentListingId = Listing::where([['board_id', $boardId], ['index', $data['current_listing_index']]])->first()->id;
+            $card->update([
+                'index' => $data['current_index'],
+                'listing_id' => $currentListingId,
+            ]);
+            if($previousListingId)
+                CardService::resetIndexesAfterAction($previousListingId);
+            CardService::resetIndexesAfterAction($currentListingId);
+            DB::commit();
             return $card;
-        }catch(Exception $e){
+        }
+        catch(AuthorizationException $e){
+            return response()->json(['message' => 'Forbidden!'], 403);
+        }
+        catch(Exception $e){
+            DB::rollBack();
             return ['message' => 'error card could not be moved!'];
         }
     }
@@ -115,15 +125,20 @@ class CardController extends Controller
     public function destroy($id)
     {
         try{
-            $card = null;
-            DB::transaction(function() use(&$card, $id){
-                $card = Card::findOrFail($id);
-                $card->delete();
-                CardService::resetIndexesAfterAction($card->listing_id);
-            });
+            DB::beginTransaction();
+            $card = Card::findOrFail($id);
+            $this->authorize('delete', [$card, $card->listing_id]);
+            $card->delete();
+            CardService::resetIndexesAfterAction($card->listing_id);
+            DB::commit();
             return $card;
-        }catch(Exception $e){
-            return ['message' => 'card could not be deleted!'];
+        }
+        catch(AuthorizationException $e){
+            return response()->json(['message' => 'Forbidden!'], 403);
+        }
+        catch(Exception $e){
+            DB::rollBack();
+            return ['message' => 'Something went wrong, card could not be deleted!'];
         }
     }
 }
